Force --no-as-needed linker flag on Linux.
Some Linux distributions (notably Ubuntu) sets the --as-needed linker
flag by default, which makes the linker sensitive of the order of the
libraries the executable file depends on in the command line. Other
distributions (such as ArchLinux) do not set this flag by default. For
safety, this flag is explicitly turned off in both the Makefile and the
refimpl2-config script on Linux.